fix: qualify exact release package evidence (#1093) (#1102)
* fix: qualify exact release package evidence

* fix: verify ignored release cases and retained attempts

* fix: prove packaged release evidence controls in PR CI

* fix: download proof package into its expected directory

* fix: preserve CDC function ownership during extension updates

* fix: repair CDC rebuild lookup and Semgrep findings

* fix: complete Semgrep suppressions

* fix: preserve defining query typmods

* fix: rebuild stream storage after schema reinitialization

* test: account for dimensioned vector averages

* test: account for dimensioned vector sums

* test: qualify vector average schema

* test: allow WAL refresh startup latency

* test: close release evidence proof gaps

* test: close release evidence proof gaps

+++ tests/e2e_pipeline_dag_tests.rs
@@ -333,8 +333,8 @@ async fn test_nx_pipeline_insert_new_bid_cascades() {
     db.assert_st_matches_query(
         "nx_auction_bids",
         "SELECT b.id AS bid_id, b.auction_id, b.bidder_id,
-                b.price::NUMERIC AS price, a.seller_id, a.category,
-                a.item_name, a.reserve_price::NUMERIC AS reserve_price
+                b.price AS price, a.seller_id, a.category,
+                a.item_name, a.reserve_price AS reserve_price
          FROM nx_bids b
          JOIN nx_auctions a ON a.id = b.auction_id",
     )
@@ -508,8 +508,8 @@ async fn test_nx_pipeline_update_auction_category_cascades() {
     db.assert_st_matches_query(
         "nx_auction_bids",
         "SELECT b.id AS bid_id, b.auction_id, b.bidder_id,
-                b.price::NUMERIC AS price, a.seller_id, a.category,
-                a.item_name, a.reserve_price::NUMERIC AS reserve_price
+                b.price AS price, a.seller_id, a.category,
+                a.item_name, a.reserve_price AS reserve_price
          FROM nx_bids b
          JOIN nx_auctions a ON a.id = b.auction_id",
     )
@@ -856,7 +856,7 @@ async fn test_ec_pipeline_insert_order_cascades_all_layers() {
     db.assert_st_matches_query(
         "ec_line_details",
         "SELECT oi.id AS line_id, oi.order_id, oi.quantity,
-                oi.unit_price::NUMERIC AS unit_price,
+                oi.unit_price AS unit_price,
                 oi.quantity * oi.unit_price AS line_total,
                 ep.product_id, ep.product_name, ep.category_id, ep.category_name
          FROM ec_order_items oi
